Melhoria na solução.
