+++ app/src/main/java/com/silvioapps/githubkotlin/features/shared/views/applications/CustomApplication.kt
@@ -1,0 +1,12 @@
+package com.silvioapps.githubkotlin.features.shared.views.applications
+
+import android.app.Application
+import com.silvioapps.githubkotlin.features.shared.utils.Utils
+
+class CustomApplication : Application(){
+
+    override fun onCreate() {
+        super.onCreate()
+        Utils.fixSSLError(getApplicationContext())
+    }
+}
